Moved the nav bar to the side
