+++ BlogPoster/src/Layout.jsx
@@ -3,7 +3,7 @@ import Header from "./Header";
 
 export default function Layout() {
     return (
-        <div className="p-4 flex flex-col min-h-screen">
+        <div className="p-4 flex flex-row min-h-screen">
             <Header />
             <Outlet />
         </div>
